requirements.txt
streamlit
gspread
gspread_asyncio
google-auth
nest_asyncio
kiteconnect==4.0.0
pandas
python-dateutil==2.8.2
